add len and coma
### Additions
`Str.len()` - retrieves the length of a string
`coma(n)` - causes the program to fall into a coma for a specified amount of time n in seconds (n can be a float too, like 0.5)

### Important Note
DO NOT change the name of `coma(n)`, the programming language WILL NOT work otherwise. I don't know how or why, but you have to leave it named like that. Source: trust me bro.